Split code off into multiple files for maintanability
