+++ index.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="UTF-8">
 		<title>Pick a DHV for me</title>
-		<script src="form-handler.js" defer></script>
+		<script type="module" src="form-handler.js" defer></script>
                 <link rel="stylesheet" href="style.css">
 	</head>
 	<body>
